# src/openapi.yaml
openapi: "3.0.0"
info:
  description: ""
  version: "1.0.0"
  title: "Sekam"
  termsOfService: ""
  contact:
    email: "admin@sekam.id"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
servers:
  - url: http://34.134.145.134
    description: Production server (uses live data)
  - url: https://localhost:8080
    description: Sandbox server (uses test data)
  - url: http://localhost:8080
    description: Local server
paths:
...
  /product:
    post:
      tags:
      - "product"
      summary: "Add Product"
      description: ""
      operationId: "addProduct"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: "Success"
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "404":
          description: "Not found"
        "405":
          description: "Validation exception"
      security: 
      - ApiKey: []

    get:
      tags:
      - "product"
      summary: "Get All Product"
      description: ""
      operationId: "getProduct"
      responses:
        "200":
          description: "Success"
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "404":
          description: "Not found"
        "405":
          description: "Validation exception"
      security: 
      - ApiKey: []

    put:
      tags:
      - "product"
      summary: "Update Product"
      description: ""
      operationId: "updateProduct"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "200":
          description: "Success"
          content:
            application/json:    
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "404":
          description: "Not found"
        "405":
          description: "Validation exception"
      security: 
      - ApiKey: []
...
components:
  securitySchemes:
    ApiKey:
       type: apiKey
       name: Authorization
       in: header
  schemas:
...
    Product:
      type: "object"
      properties:
        id:
          type: "integer"
          description: "only required on update"
        name:
          type: "string"
        price:
          type: "string"
        description:
          type: "string"
        merchantId:
          type: "string"
        categoryId:
          type: "string"
      xml:
        name: "Product"
...
    ApiResponse:
      type: "object"
      properties:
        data:
          type: "object"
        message:
          type: "string"
...